Combine notification settings into main settings UI.
