+++ DuckDuckGo/NetworkProtectionVPNSettingsViewModel.swift
@@ -20,18 +20,33 @@
 #if NETWORK_PROTECTION
 
 import Foundation
+import UserNotifications
 import NetworkProtection
 import Combine
+
+enum NetworkProtectionNotificationsViewKind: Equatable {
+    case loading
+    case unauthorized
+    case authorized
+}
 
 final class NetworkProtectionVPNSettingsViewModel: ObservableObject {
     private let settings: VPNSettings
     private var cancellables: Set<AnyCancellable> = []
 
+    private var notificationsAuthorization: NotificationsAuthorizationControlling
+    @Published var viewKind: NetworkProtectionNotificationsViewKind = .loading
+
+    var alertsEnabled: Bool {
+        self.settings.notifyStatusChanges
+    }
+
     @Published public var preferredLocation: NetworkProtectionLocationSettingsItemModel
     @Published public var excludeLocalNetworks: Bool = true
 
-    init(settings: VPNSettings) {
+    init(notificationsAuthorization: NotificationsAuthorizationControlling, settings: VPNSettings) {
         self.settings = settings
+        self.notificationsAuthorization = notificationsAuthorization
         self.preferredLocation = NetworkProtectionLocationSettingsItemModel(selectedLocation: settings.selectedLocation)
         settings.selectedLocationPublisher
             .receive(on: DispatchQueue.main)
@@ -45,12 +60,37 @@ final class NetworkProtectionVPNSettingsViewModel: ObservableObject {
             .store(in: &cancellables)
     }
 
+    @MainActor
+    func onViewAppeared() async {
+        let status = await notificationsAuthorization.authorizationStatus
+        updateViewKind(for: status)
+    }
+
+    func turnOnNotifications() {
+        notificationsAuthorization.requestAlertAuthorization()
+    }
+
+    func didToggleAlerts(to enabled: Bool) {
+        settings.notifyStatusChanges = enabled
+    }
+
     func toggleExcludeLocalNetworks() {
         settings.excludeLocalNetworks.toggle()
     }
 
     private static func localizedString(forRegionCode: String) -> String {
         Locale.current.localizedString(forRegionCode: forRegionCode) ?? forRegionCode.capitalized
+    }
+
+    private func updateViewKind(for authorizationStatus: UNAuthorizationStatus) {
+        switch authorizationStatus {
+        case .notDetermined, .denied:
+            viewKind = .unauthorized
+        case .authorized, .ephemeral, .provisional:
+            viewKind = .authorized
+        @unknown default:
+            assertionFailure("Unhandled enum case")
+        }
     }
 }
 
@@ -80,6 +120,12 @@ struct NetworkProtectionLocationSettingsItemModel {
             }
             icon = .emoji(countryLabelsModel.emoji)
         }
+    }
+}
+
+extension NetworkProtectionVPNSettingsViewModel: NotificationsPermissionsControllerDelegate {
+    func authorizationStateDidChange(toStatus status: UNAuthorizationStatus) {
+        updateViewKind(for: status)
     }
 }
